pyproject.toml
[project]
name = "cccencode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "flagembedding>=1.3.4",
    "jieba>=0.42.1",
    "langchain>=0.3.22",
    "langchain-community>=0.3.20",
    "langchain-ollama>=0.3.0",
    "modelscope>=1.25.0",
    "openai>=1.70.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pip>=25.0.1",
    "rank-bm25>=0.2.2",
    "streamlit==1.32.2",
    "torch==2.6.0",
    "transformers>=4.50.3",
    "watchdog>=6.0.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true